Split launcher specific code from DataFilesList back to DataFilesPage.
